+++ memo/src/App.tsx
@@ -1,30 +1,23 @@
 import { ChangeEvent, useState, FC, useCallback } from "react";
 import styled from "styled-components";
 import { MemoList } from "./MemoList";
+import { useMemoList } from "./hooks/useMemoList";
 
 export const App: FC = () => {
+  const {memos, addMemo, deleteMemo} = useMemoList();
   const [text, setText] = useState<string>("");
-
-  const [memos, setMemos] = useState<string[]>([]);
 
   const onChangeText = (e: ChangeEvent<HTMLInputElement>) =>
     setText(e.target.value);
 
   const onClickAdd = () => {
-    const newMemos = [...memos];
-    newMemos.push(text);
-    setMemos(newMemos);
+    addMemo(text);
     setText("");
   };
 
-  const onClickDelete = useCallback(
-    (index: number) => {
-      const newMemos = [...memos];
-      newMemos.splice(index, 1);
-      setMemos(newMemos);
-    },
-    [memos],
-  );
+  const onClickDelete = useCallback((index: number) => {
+    deleteMemo(index);
+    }, [deleteMemo]);
 
   return (
     <div>
